requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.27
Werkzeug==3.0.1
email-validator==2.1.0.post1
bcrypt==4.1.2
pyotp==2.9.0
python-dotenv==1.0.1
Flask-Principal==0.4.0
python-dateutil==2.8.2
pytz==2024.1 